Update to 3.5.0 (#223)
- Create translated resource children by option
- Refresh translated resources with the content of a source resource
- Mass unlink/delete/refresh translated resources of one resource and their children by option
- Hide buttons in the custom manager page for contexts not in the context group of the current resource - thanks to A Demyanenko <https://github.com/rs-a-demyanenko>
- Enable/disable babel.syncFields and babel.syncTVs with context settings - thanks to A Demyanenko <https://github.com/rs-a-demyanenko>
- Sync babel.syncFields and babel.syncTVs context settings in grouped contexts - thanks to A Demyanenko <https://github.com/rs-a-demyanenko>
- The babel button menu will have a scroll bar, with a large count of contexts
- Add a new system setting `babel.displayChunk`, that can define a chunk for `babel.displayText` system setting

+++ source/js/mgr/widgets/resourcematrix.grid.js
@@ -44,7 +44,7 @@ Babel.grid.ResourceMatrix = function (config) {
             menuDisabled: true,
             fixed: true,
             locked: true,
-            width: 50
+            width: 85
         }];
         Ext.each(config.contexts, function (item) {
             fields.push('linkedres_id_' + item);
@@ -52,7 +52,7 @@ Babel.grid.ResourceMatrix = function (config) {
             contexts.push(item);
             columns.push({
                 header: item,
-                width: 70,
+                width: 85,
                 sortable: false,
                 dataIndex: 'linkedres_id_' + item,
                 id: 'linkedres_id_' + item,
@@ -131,11 +131,17 @@ Ext.extend(Babel.grid.ResourceMatrix, MODx.grid.Grid, {
     linkTranslation: function (ctx, id) {
         Babel.linkTranslation(ctx, id, this);
     },
-    unlinkTranslation: function (ctx, id, target) {
-        return Babel.unlinkTranslation(ctx, id, target, this);
+    unlinkTranslation: function (ctx, id) {
+        return Babel.unlinkTranslation(ctx, id, this);
+    },
+    deleteTranslation: function (ctx, id) {
+        return Babel.deleteTranslation(ctx, id, this);
     },
     createTranslation: function (ctx, id) {
         Babel.createTranslation(ctx, id)
+    },
+    refreshTranslation: function (ctx, id) {
+        Babel.refreshTranslation(ctx, id)
     },
     clearFilter: function () {
         var store = this.getStore();
@@ -164,9 +170,17 @@ Ext.extend(Babel.grid.ResourceMatrix, MODx.grid.Grid, {
                 icon: 'plus-square',
                 text: _('babel.create_multiple_translations')
             }, {
+                className: 'refresh-multiple',
+                icon: 'refresh',
+                text: _('babel.refresh_multiple_translations')
+            }, {
                 className: 'unlink-all',
                 icon: 'chain-broken',
                 text: _('babel.unlink_all')
+            }, {
+                className: 'delete-all',
+                icon: 'trash',
+                text: _('babel.delete_all')
             }]
         });
     },
@@ -174,36 +188,46 @@ Ext.extend(Babel.grid.ResourceMatrix, MODx.grid.Grid, {
         var actionButtons = [];
         if (metaData.id !== 'linkedres_id_' + record.get('context_key')) {
             var ctx = metaData.id.substr('linkedres_id_'.length);
-            if (record.get(metaData.id) === '') {
-                actionButtons.push({
-                    className: 'create',
-                    icon: 'plus-circle',
-                    text: _('babel.create_translation'),
-                    ctx: ctx,
-                    target: 0
-                }, {
-                    className: 'link',
-                    icon: 'link',
-                    text: _('babel.link_translation'),
-                    ctx: ctx,
-                    target: 0
-                });
-            } else {
-                var pagetitle = record.get('linkedres_pagetitle_' + ctx);
-                var target = record.get('linkedres_id_' + ctx);
-                actionButtons.push({
-                    className: 'update',
-                    icon: 'pencil-square-o',
-                    text: _('edit') + ': ' + pagetitle + ' (' + target + ')',
-                    ctx: ctx,
-                    target: target
-                }, {
-                    className: 'unlink',
-                    icon: 'chain-broken',
-                    text: _('babel.unlink') + ': ' + pagetitle + ' (' + target + ')',
-                    ctx: ctx,
-                    target: target
-                });
+            var target = record.get('linkedres_id_' + ctx);
+            if (target !== 'x') {
+                if (record.get(metaData.id) === '') {
+                    actionButtons.push({
+                        className: 'create',
+                        icon: 'plus-circle',
+                        text: _('babel.create_translation'),
+                        ctx: ctx,
+                    }, {
+                        className: 'link',
+                        icon: 'link',
+                        text: _('babel.link_translation'),
+                        ctx: ctx,
+                    });
+                } else {
+                    var pagetitle = record.get('linkedres_pagetitle_' + ctx);
+                    actionButtons.push({
+                        className: 'update',
+                        icon: 'pencil-square-o',
+                        text: _('edit') + ': ' + pagetitle + ' (' + target + ')',
+                        ctx: ctx,
+                        target: target
+                    }, {
+                        className: 'refresh',
+                        icon: 'refresh',
+                        text: _('babel.refresh') + ': ' + pagetitle + ' (' + target + ')',
+                        ctx: ctx,
+                        target: target
+                    }, {
+                        className: 'unlink',
+                        icon: 'chain-broken',
+                        text: _('babel.unlink') + ': ' + pagetitle + ' (' + target + ')',
+                        ctx: ctx,
+                    }, {
+                        className: 'delete',
+                        icon: 'trash-o',
+                        text: _('babel.delete') + ': ' + pagetitle + ' (' + target + ')',
+                        ctx: ctx,
+                    });
+                }
             }
         }
         return this.buttonColumnTpl.apply({
@@ -217,14 +241,11 @@ Ext.extend(Babel.grid.ResourceMatrix, MODx.grid.Grid, {
             var act = t.className.split(' ')[1];
             var record = this.getSelectionModel().getSelected();
             switch (act) {
-                case 'unlink-all':
-                    this.unlinkTranslation('', record.get('id'), 0)
+                case 'create':
+                    this.createTranslation(t.dataset.ctx, record.get('id'));
                     break;
                 case 'create-multiple':
                     this.createTranslation('', record.get('id'));
-                    break;
-                case 'create':
-                    this.createTranslation(t.dataset.ctx, record.get('id'));
                     break;
                 case 'link':
                     this.linkTranslation(t.dataset.ctx, record.get('id'));
@@ -232,8 +253,23 @@ Ext.extend(Babel.grid.ResourceMatrix, MODx.grid.Grid, {
                 case 'update':
                     MODx.loadPage('resource/update', 'id=' + t.dataset.target);
                     break;
+                case 'refresh':
+                    this.refreshTranslation(t.dataset.ctx, record.get('id'));
+                    break;
+                case 'refresh-multiple':
+                    this.refreshTranslation('', record.get('id'));
+                    break;
                 case 'unlink':
-                    this.unlinkTranslation(t.dataset.ctx, record.get('id'), t.dataset.target);
+                    this.unlinkTranslation(t.dataset.ctx, record.get('id'));
+                    break;
+                case 'unlink-all':
+                    this.unlinkTranslation('', record.get('id'), 0);
+                    break;
+                case 'delete':
+                    this.deleteTranslation(t.dataset.ctx, record.get('id'));
+                    break;
+                case 'delete-all':
+                    this.deleteTranslation('', record.get('id'), 0);
                     break;
                 default:
                     break;
